# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TrackVault API",
        "description": "Data Collection and Payment Management System - Complete REST API Documentation",
        "contact": {
            "name": "TrackVault Support",
            "email": "admin@trackvault.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Development Server"
        },
        {
            "url": "https://api.trackvault.com/api",
            "description": "Production Server"
        }
    ],
    "paths": {
...
        "/api/collections": {
            "get": {
                "tags": [
                    "Collections"
                ],
                "summary": "List all collections",
                "description": "Get paginated list of collections with filtering and sorting",
                "operationId": "cd6db30c7ea18bad400a6fcc1902ed6a",
                "parameters": [
                    {
                        "name": "supplier_id",
                        "in": "query",
                        "description": "Filter by supplier ID",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "product_id",
                        "in": "query",
                        "description": "Filter by product ID",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "from_date",
                        "in": "query",
                        "description": "Filter from date (YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "to_date",
                        "in": "query",
                        "description": "Filter to date (YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "sort_by",
                        "in": "query",
                        "description": "Sort by field",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "collection_date",
                            "enum": [
                                "collection_date",
                                "quantity",
                                "total_amount",
                                "created_at",
                                "updated_at"
                            ]
                        }
                    },
                    {
                        "name": "sort_order",
                        "in": "query",
                        "description": "Sort order",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "desc",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Results per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15,
                            "maximum": 100
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "supplier_id": {
                                                        "type": "integer"
                                                    },
                                                    "product_id": {
                                                        "type": "integer"
                                                    },
                                                    "collection_date": {
                                                        "type": "string",
                                                        "format": "date"
                                                    },
                                                    "quantity": {
                                                        "type": "number"
                                                    },
                                                    "unit": {
                                                        "type": "string"
                                                    },
                                                    "rate_applied": {
                                                        "type": "number"
                                                    },
                                                    "total_amount": {
                                                        "type": "number"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Collections"
                ],
                "summary": "Create a new collection",
                "description": "Create a new collection record. Rate and total amount are calculated automatically.",
                "operationId": "03cd1f353d77b52603ebd7680a4a8b85",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "supplier_id",
                                    "product_id",
                                    "collection_date",
                                    "quantity",
                                    "unit"
                                ],
                                "properties": {
                                    "supplier_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "product_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "collection_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-12-25"
                                    },
                                    "quantity": {
                                        "type": "number",
                                        "minimum": 0.001,
                                        "example": 50.5
                                    },
                                    "unit": {
                                        "type": "string",
                                        "maxLength": 50,
                                        "example": "kg"
                                    },
                                    "notes": {
                                        "type": "string",
                                        "example": "Afternoon collection"
                                    },
                                    "metadata": {
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Collection created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "supplier_id": {
                                            "type": "integer"
                                        },
                                        "product_id": {
                                            "type": "integer"
                                        },
                                        "collection_date": {
                                            "type": "string",
                                            "format": "date"
                                        },
                                        "quantity": {
                                            "type": "number"
                                        },
                                        "unit": {
                                            "type": "string"
                                        },
                                        "rate_applied": {
                                            "description": "Automatically applied rate",
                                            "type": "number"
                                        },
                                        "total_amount": {
                                            "description": "Calculated as quantity * rate_applied",
                                            "type": "number"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter your bearer token in the format: Bearer {token}",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}